Ready to submit

+++ Project3/PiratesVNinjas/normDist.c
@@ -3,7 +3,7 @@
 #include "normDist.h"
 
 double norm_dist(double mean, double stddev) {
-    int v = -1;
+    double v = -1;
     while (v <= 0) {
         v = (sqrt(-2.0 * log(drand48())) * cos(2.0 * PI * drand48())) * stddev + mean;
     }
